Move form section
