+++ app/components/base/form/create-form/index.ts
@@ -1,2 +1,1 @@
 export * from "./create-form.component";
-export * from "./form-section.component";
