+++ app/components/base/form/form-section/index.ts
@@ -1,0 +1,1 @@
+export * from "./form-section.component";
